Changed int and uint_8 to use byte, because int is using unnecessary memory

+++ Experimental/Liquid-Crystal-Custom/src/CustomLiquidCrystal.cpp
@@ -1,12 +1,12 @@
 #include "CustomLiquidCrystal.h"
 
 /*TODO: Default constructor to move same initialization in it and call it from these?!*/
-CustomLiquidCrystal::CustomLiquidCrystal(int registerSyncPinNumber,
-                                         int enablePinNumber,
-                                         int dataBus7PinNumber,
-                                         int dataBus6PinNumber,
-                                         int dataBus5PinNumber,
-                                         int dataBus4PinNumber) : registerSyncPinNumber(registerSyncPinNumber),
+CustomLiquidCrystal::CustomLiquidCrystal(byte registerSyncPinNumber,
+                                         byte enablePinNumber,
+                                         byte dataBus7PinNumber,
+                                         byte dataBus6PinNumber,
+                                         byte dataBus5PinNumber,
+                                         byte dataBus4PinNumber) : registerSyncPinNumber(registerSyncPinNumber),
                                                                   enablePinNumber(enablePinNumber),
                                                                   dataBus7PinNumber(dataBus7PinNumber),
                                                                   dataBus6PinNumber(dataBus6PinNumber),
@@ -24,16 +24,16 @@ CustomLiquidCrystal::CustomLiquidCrystal(int registerSyncPinNumber,
     this->cursorBlink = CursorBlink::OFF;
 }
 
-CustomLiquidCrystal::CustomLiquidCrystal(int registerSyncPinNumber,
-                                         int enablePinNumber,
-                                         int dataBus7PinNumber,
-                                         int dataBus6PinNumber,
-                                         int dataBus5PinNumber,
-                                         int dataBus4PinNumber,
-                                         int dataBus3PinNumber,
-                                         int dataBus2PinNumber,
-                                         int dataBus1PinNumber,
-                                         int dataBus0PinNumber) : registerSyncPinNumber(registerSyncPinNumber),
+CustomLiquidCrystal::CustomLiquidCrystal(byte registerSyncPinNumber,
+                                         byte enablePinNumber,
+                                         byte dataBus7PinNumber,
+                                         byte dataBus6PinNumber,
+                                         byte dataBus5PinNumber,
+                                         byte dataBus4PinNumber,
+                                         byte dataBus3PinNumber,
+                                         byte dataBus2PinNumber,
+                                         byte dataBus1PinNumber,
+                                         byte dataBus0PinNumber) : registerSyncPinNumber(registerSyncPinNumber),
                                                                   enablePinNumber(enablePinNumber),
                                                                   dataBus7PinNumber(dataBus7PinNumber),
                                                                   dataBus6PinNumber(dataBus6PinNumber),
@@ -61,10 +61,10 @@ CustomLiquidCrystal::CustomLiquidCrystal(int registerSyncPinNumber,
 
 /*TODO: More abstract way*/
 void CustomLiquidCrystal::send(RegisterSelect registerSelect,
-                               int dataBus7PinValue,
-                               int dataBus6PinValue,
-                               int dataBus5PinValue,
-                               int dataBus4PinValue) {
+                               byte dataBus7PinValue,
+                               byte dataBus6PinValue,
+                               byte dataBus5PinValue,
+                               byte dataBus4PinValue) {
 
     /*TODO: Prepare bus lines, set rs, enable, disable*/
     digitalWrite(this->dataBus7PinNumber, dataBus7PinValue);
@@ -72,7 +72,7 @@ void CustomLiquidCrystal::send(RegisterSelect registerSelect,
     digitalWrite(this->dataBus5PinNumber, dataBus5PinValue);
     digitalWrite(this->dataBus4PinNumber, dataBus4PinValue);
 
-    digitalWrite(this->registerSyncPinNumber, static_cast<uint8_t>(registerSelect)); // RS Type
+    digitalWrite(this->registerSyncPinNumber, static_cast<byte>(registerSelect)); // RS Type
     digitalWrite(this->enablePinNumber, HIGH); // Set receiving on.
     delay(1);
     digitalWrite(this->enablePinNumber, LOW); // Set receiving off, pushing everything that came after on
@@ -80,11 +80,11 @@ void CustomLiquidCrystal::send(RegisterSelect registerSelect,
 }
 
 template<typename T, size_t N>
-void CustomLiquidCrystal::getBits(uint8_t value, T (& bits)[N]) {
+void CustomLiquidCrystal::getBits(byte value, T (& bits)[N]) {
 
     int reverseCounterIndex = N - 1;
     while (value != 0) {
-        uint8_t currentBit = value & 1;
+        byte currentBit = value & 1;
 
         bits[reverseCounterIndex] = currentBit;
         reverseCounterIndex--;
@@ -92,9 +92,9 @@ void CustomLiquidCrystal::getBits(uint8_t value, T (& bits)[N]) {
     }
 }
 
-void CustomLiquidCrystal::send(RegisterSelect registerSelect, int dataBusBits) {
+void CustomLiquidCrystal::send(RegisterSelect registerSelect, byte dataBusBits) {
 
-    uint8_t bits[8] = {0, 0, 0, 0, 0, 0, 0, 0};
+    byte bits[8] = {0, 0, 0, 0, 0, 0, 0, 0};
     this->getBits(dataBusBits, bits);
 
     this->send(registerSelect, bits[0], bits[1], bits[2], bits[3]);
@@ -127,11 +127,11 @@ void CustomLiquidCrystal::set4BitInterface() {
 }
 
 void CustomLiquidCrystal::setFunction(InterfaceLength interfaceLength, Lines lines, CharacterFont characterFont) {
-    uint8_t interfaceLengthMask = static_cast<uint8_t>(interfaceLength) << 4;
-    uint8_t linesMask = static_cast<uint8_t>(lines) << 3;
-    uint8_t characterFontMask = static_cast<uint8_t>(characterFont) << 2;
+    byte interfaceLengthMask = static_cast<byte>(interfaceLength) << 4;
+    byte linesMask = static_cast<byte>(lines) << 3;
+    byte characterFontMask = static_cast<byte>(characterFont) << 2;
 
-    uint8_t commandBits = 0b00100000 | interfaceLengthMask | linesMask | characterFontMask;
+    byte commandBits = 0b00100000 | interfaceLengthMask | linesMask | characterFontMask;
     this->send(RegisterSelect::COMMAND, commandBits);
 }
 
@@ -141,10 +141,10 @@ void CustomLiquidCrystal::clearDisplay() {
 
 void CustomLiquidCrystal::setEntryMode(CursorDirection cursorDirection, DisplayShift displayShift) {
 
-    uint8_t cursorDirectionMask = static_cast<uint8_t>(cursorDirection) << 1;
-    uint8_t displayShiftMask = static_cast<uint8_t>(displayShift) << 0;
+    byte cursorDirectionMask = static_cast<byte>(cursorDirection) << 1;
+    byte displayShiftMask = static_cast<byte>(displayShift) << 0;
 
-    uint8_t commandBits = 0b00000100 | cursorDirectionMask | displayShiftMask;
+    byte commandBits = 0b00000100 | cursorDirectionMask | displayShiftMask;
     this->send(RegisterSelect::COMMAND, commandBits);
 }
 
@@ -161,11 +161,11 @@ void CustomLiquidCrystal::blinkCursor(CursorBlink cursorBlink) {
 }
 
 void CustomLiquidCrystal::setDisplay(Display display, CursorToggle cursorToggle, CursorBlink cursorBlink) {
-    uint8_t displayMask = static_cast<uint8_t>(display) << 2;
-    uint8_t cursorToggleMask = static_cast<uint8_t>(cursorToggle) << 1;
-    uint8_t cursorBlinkMask = static_cast<uint8_t>(cursorBlink) << 0;
+    byte displayMask = static_cast<byte>(display) << 2;
+    byte cursorToggleMask = static_cast<byte>(cursorToggle) << 1;
+    byte cursorBlinkMask = static_cast<byte>(cursorBlink) << 0;
 
-    uint8_t commandBits = 0b00001000 | displayMask | cursorToggleMask | cursorBlinkMask;
+    byte commandBits = 0b00001000 | displayMask | cursorToggleMask | cursorBlinkMask;
     this->send(RegisterSelect::COMMAND, commandBits);
 
     this->display = display;
